+++ src/plugins/paging/basic/basicPaging.component.ts
@@ -125,6 +125,14 @@ export class BasicPagingComponent extends PagingAbstractComponent<CssClassesBasi
         return isNaN(offset) ? 0 : offset;
     }
 
+    /**
+     * Indication whether plugin is already initialized
+     */
+    public get initialized(): boolean
+    {
+        return this._initialized;
+    }
+
     //######################### public properties - inputs #########################
 
     /**